feat: group the component
