+++ at_onboarding_flutter/lib/at_onboarding_pair_screen.dart
@@ -2,8 +2,8 @@ import 'dart:convert';
 
 import 'package:at_onboarding_flutter/services/size_config.dart';
 import 'package:at_onboarding_flutter/utils/at_onboarding_dimens.dart';
+import 'package:at_onboarding_flutter/utils/at_onboarding_error_util.dart';
 import 'package:at_onboarding_flutter/utils/color_constants.dart';
-import 'package:at_onboarding_flutter/utils/error_util.dart';
 import 'package:at_onboarding_flutter/utils/strings.dart';
 import 'package:at_onboarding_flutter/widgets/at_onboarding_button.dart';
 import 'package:flutter/gestures.dart';
@@ -380,7 +380,7 @@ class _AtOnboardingPairScreenState extends State<AtOnboardingPairScreen> {
   }
 
   Future<void> _showAlertDialog(dynamic errorMessage, {String? title}) async {
-    String? messageString = ConvertErrorToString().getErrorMessage(errorMessage);
+    String? messageString = AtOnboardingErrorToString().getErrorMessage(errorMessage);
     return AtOnboardingDialog.showError(
         context: context, message: messageString);
   }
