--- at_onboarding_flutter/lib/utils/images.dart
@@ -1,4 +1,0 @@
-class Images {
-  static const String _basepath = 'assets/images';
-  static const String backupZip = '$_basepath/backup_key.png';
-}
